to tpl that eventbus optimized
