+++ app/build.gradle
@@ -150,7 +150,6 @@ dependencies {
     implementation "com.facebook.stetho:stetho-okhttp3:$rootProject.stetho_okhttp3_version"
     implementation "com.facebook.stetho:stetho-js-rhino:$rootProject.stetho_js_rhino_version"
     implementation "com.orhanobut:logger:$rootProject.logger_version"
-    implementation "org.greenrobot:eventbus:$rootProject.eventbus_version"
     implementation "junit:junit:$rootProject.junit_version"
     testImplementation "com.squareup.leakcanary:leakcanary-android-no-op:$rootProject.leakcanary_android_no_op_version"
     debugImplementation "com.squareup.leakcanary:leakcanary-android:$rootProject.leakcanary_android"
