+++ app/src/main/java/com/dvsnier/wrapper/TransferStationWrapper.java
@@ -25,7 +25,7 @@ import com.dvsnier.test.widget.recycleview.TestRecyclerActivity;
 import com.dvsnier.test.widget.scroll.TestScrollActivity;
 import com.dvsnier.test.widget.surfaceview.TestSurfaceActivity;
 import com.dvsnier.test.widget.theme.TestThemeActivity;
-import com.dvsnier.testEventBus.TestEventBusActivity;
+import com.dvsnier.test.tpl.eventbus.TestEventBusActivity;
 import com.dvsnier.test.tpl.green.TestGreenDaoActivity;
 import com.dvsnier.test.tpl.okhttp.TestOkhttpActivity;
 import com.dvsnier.test.tpl.speech.TestSpeechRecognitionActivity;
