+++ framework/base/src/main/java/com/dvsnier/AbstractBaseApplication.java
@@ -1,8 +1,10 @@
 package com.dvsnier;
 
+import android.app.Application;
+
 /**
  * AbstractBaseApplication
  * Created by dovsnier on 2019/6/26.
  */
-public abstract class AbstractBaseApplication<T> extends BaseApplication implements IBaseApplication<T> {
+public abstract class AbstractBaseApplication<T> extends Application implements IBaseApplication<T> {
 }
