+++ framework/base/src/main/java/com/dvsnier/BaseApplication.java
@@ -1,10 +1,8 @@
 package com.dvsnier;
 
-import android.app.Application;
-
 /**
  * BaseApplication
  * Created by dovsnier on 2019/6/25.
  */
-public class BaseApplication extends Application implements IApplication {
+public abstract class BaseApplication<T> extends AbstractBaseApplication<T> {
 }
