+++ framework/base/src/main/java/com/dvsnier/IInstance.java
@@ -1,0 +1,12 @@
+package com.dvsnier;
+
+/**
+ * IInstance
+ * Created by dovsnier on 2019/6/26.
+ *
+ * @deprecated
+ */
+public interface IInstance<T> extends Instance {
+
+    T getInstance();
+}
