+++ framework/base/src/main/java/com/dvsnier/Instance.java
@@ -1,0 +1,8 @@
+package com.dvsnier;
+
+/**
+ * Instance
+ * Created by dovsnier on 2019/6/26.
+ */
+public interface Instance {
+}
