+++ research/tpl/src/main/AndroidManifest.xml
@@ -26,5 +26,9 @@
             android:label="@string/title_activity_test_okhttp"
             android:theme="@style/AppTheme.DefaultTheme"
             android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
+        <activity
+            android:name=".eventbus.TestEventBusActivity"
+            android:theme="@style/AppTheme.DefaultTheme"
+            android:windowSoftInputMode="adjustNothing|stateAlwaysHidden" />
     </application>
 </manifest>
