+++ research/tpl/src/main/res/layout/activity_test_event_bus.xml
